refactor(agent): move activity_log under agent module
