+++ src/agent/activity_log.rs
