+++ src/agent/manager.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 use tokio::sync::Mutex;
 use tracing::{debug, error, info, warn};
 
-use crate::activity_log::{self, ActivityEntry, ActivityLogMap, ActivityLogResponse};
+use crate::agent::activity_log::{self, ActivityEntry, ActivityLogMap, ActivityLogResponse};
 use crate::agent::drivers::{Driver, ParsedEvent, SpawnContext};
 use crate::server::AgentLifecycle;
 use crate::store::agents::{AgentConfig, AgentStatus};
