+++ src/agent/mod.rs
@@ -1,3 +1,4 @@
+pub mod activity_log;
 pub mod drivers;
 pub mod manager;
 pub mod workspace;
