+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod activity_log;
 pub mod agent;
 pub mod bridge;
 pub mod server;
