+++ src/server/handlers/agents.rs
@@ -7,7 +7,7 @@ use serde::Deserialize;
 use tracing::info;
 
 use super::{acquire_transition, api_err, internal_err, ApiResult, AppState};
-use crate::activity_log::ActivityLogResponse;
+use crate::agent::activity_log::ActivityLogResponse;
 use crate::agent::workspace::AgentWorkspace;
 use crate::store::agents::{Agent, AgentEnvVar, AgentStatus};
 use crate::store::messages::SenderType;
