+++ src/server/handlers/messages.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 use tracing::{debug, info};
 
 use super::{api_err, internal_err, ApiResult, AppState};
-use crate::activity_log::ActivityEntry;
+use crate::agent::activity_log::ActivityEntry;
 use crate::store::agents::AgentStatus;
 use crate::store::channels::ChannelType;
 use crate::store::messages::{ReceivedMessage, SenderType};
