+++ src/server/mod.rs
@@ -10,7 +10,7 @@ use axum::Router;
 use tower_http::cors::{Any, CorsLayer};
 use tower_http::services::{ServeDir, ServeFile};
 
-use crate::activity_log::{ActivityEntry, ActivityLogResponse};
+use crate::agent::activity_log::{ActivityEntry, ActivityLogResponse};
 use crate::store::messages::ReceivedMessage;
 use crate::store::Store;
